Tidy up FullApplicationTest
This is the base class for *all* full application tests, so it should be
minimal and only contain stuff needed for the majority of tests.
