+++ service/src/integrationTest/kotlin/fi/espoo/evaka/varda/VardaClientIntegrationTest.kt
@@ -5,7 +5,6 @@
 package fi.espoo.evaka.varda
 
 import com.github.kittinunf.fuel.core.isSuccessful
-import fi.espoo.evaka.FullApplicationTest
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
@@ -13,7 +12,7 @@ import org.springframework.test.context.TestPropertySource
 import kotlin.test.assertTrue
 
 @TestPropertySource(properties = ["fi.espoo.voltti.vtj.xroad.trustStore.location=classpath:test-certificate/localhost.truststore"])
-class VardaClientIntegrationTest : FullApplicationTest() {
+class VardaClientIntegrationTest : VardaIntegrationTest() {
     private lateinit var httpsServer: VardaClientIntegrationMockHttpsServer
 
     @BeforeAll
